# obfuscate: s/strings.Builder/bytes.Buffer (downgrade to go1.9.4)