--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -5,7 +5,7 @@ jobs:
           resource_class: large
 
           docker:
-               - image: circleci/golang:1.10
+               - image: circleci/golang:1.9.4
 
           steps:
                - checkout