--- a/obfuscate/json.go
+++ b/obfuscate/json.go
@@ -1,6 +1,7 @@
 package obfuscate
 
 import (
+	"bytes"
 	"strings"
 
 	"github.com/DataDog/datadog-trace-agent/config"
@@ -54,7 +55,7 @@ func (p *jsonObfuscator) setKey() {
 }
 
 func (p *jsonObfuscator) obfuscate(data []byte) (string, error) {
-	var out strings.Builder
+	var out bytes.Buffer
 	buf := make([]byte, 0, 10) // recording key token
 	p.scan.reset()
 	for _, c := range data {