A small performance hit is observed:
```
name                  old time/op    new time/op    delta
ObfuscateJSON/5003-8    48.4µs ± 0%    46.9µs ± 0%   -3.10%
ObfuscateJSON/437-8     6.77µs ± 0%    6.64µs ± 0%   -1.93%
ObfuscateJSON/671-8     6.81µs ± 0%    6.86µs ± 0%   +0.73%

name                  old alloc/op   new alloc/op   delta
ObfuscateJSON/5003-8    22.1kB ± 0%    23.8kB ± 0%   +7.41%
ObfuscateJSON/437-8     2.95kB ± 0%    3.47kB ± 0%  +17.62%
ObfuscateJSON/671-8     3.02kB ± 0%    3.47kB ± 0%  +15.12%

name                  old allocs/op  new allocs/op  delta
ObfuscateJSON/5003-8       203 ± 0%       199 ± 0%   -1.97%
ObfuscateJSON/437-8       48.0 ± 0%      46.0 ± 0%   -4.17%
ObfuscateJSON/671-8       42.0 ± 0%      40.0 ± 0%   -4.76%
```